+++ packages/ai/CHANGELOG.md
@@ -1,3 +1,7 @@
+## @lunora/ai [1.0.0-alpha.7](https://github.com/anolilab/lunora/compare/%40lunora%2Fai%401.0.0-alpha.6...%40lunora%2Fai%401.0.0-alpha.7) (2026-07-04)
+
+## @lunora/ai [1.0.0-alpha.6](https://github.com/anolilab/lunora/compare/%40lunora%2Fai%401.0.0-alpha.5...%40lunora%2Fai%401.0.0-alpha.6) (2026-07-04)
+
 ## @lunora/ai [1.0.0-alpha.5](https://github.com/anolilab/lunora/compare/%40lunora%2Fai%401.0.0-alpha.4...%40lunora%2Fai%401.0.0-alpha.5) (2026-07-04)
 
 ## @lunora/ai [1.0.0-alpha.4](https://github.com/anolilab/lunora/compare/%40lunora%2Fai%401.0.0-alpha.3...%40lunora%2Fai%401.0.0-alpha.4) (2026-07-03)