+++ packages/db/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/db [1.0.0-alpha.18](https://github.com/anolilab/lunora/compare/%40lunora%2Fdb%401.0.0-alpha.17...%40lunora%2Fdb%401.0.0-alpha.18) (2026-07-04)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.18
+
 ## @lunora/db [1.0.0-alpha.17](https://github.com/anolilab/lunora/compare/%40lunora%2Fdb%401.0.0-alpha.16...%40lunora%2Fdb%401.0.0-alpha.17) (2026-07-03)