+++ packages/studio/CHANGELOG.md
@@ -1,3 +1,19 @@
+## @lunora/studio [1.0.0-alpha.38](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.37...%40lunora%2Fstudio%401.0.0-alpha.38) (2026-07-04)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.18
+* **@lunora/react:** upgraded to 1.0.0-alpha.20
+* **@lunora/mail:** upgraded to 1.0.0-alpha.9
+
+## @lunora/studio [1.0.0-alpha.37](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.36...%40lunora%2Fstudio%401.0.0-alpha.37) (2026-07-04)
+
+
+### Dependencies
+
+* **@lunora/mail:** upgraded to 1.0.0-alpha.8
+
 ## @lunora/studio [1.0.0-alpha.36](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.35...%40lunora%2Fstudio%401.0.0-alpha.36) (2026-07-04)